% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtweetxtras_functions.R
\name{bar_plot_mentions}
\alias{bar_plot_mentions}
\title{Twitter Mentions Barplot for rtweet Package}
\usage{
bar_plot_mentions(rtweet_df, no_of_bars = 20, title = NULL)
}
\arguments{
\item{rtweet_df}{A dataframe/tibble of tweets created by rtweet package through twitter API call.}

\item{no_of_bars}{The number of bars in the plot. Default is 20.}

\item{title}{The title of the barplot in quotation marks. Default of NULL returns the name of the rtweet tibble and the number of bars.}
}
\description{
This function creates a barplot of mentioned screen_names (default top 20) in an rtweet tibble of tweets
}
\examples{
df <- rtweet::get_timelines("jack", n=3200)
bar_plot_mentions(df, no_of_bars = 25, title = "Jack's Mentions")
}
\keyword{mentions}
\keyword{rtweet,}
\keyword{twitter,}
\keyword{visualization,}
