% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkshadowban.R
\name{check_shadowban}
\alias{check_shadowban}
\title{Shadowban check}
\usage{
check_shadowban(screen_name, timezone = "UTC")
}
\arguments{
\item{screen_name}{A twitter screen_name (case insensitive)}

\item{timezone}{Optional, a timezone for the timestamp of the ban check. Default is "UTC"}
}
\description{
Checks a particular twitter screen_name for temporary search or reply visibility reduction.
}
\examples{
check <- check_shadowban("jack", "Africa/Cairo")
}
\keyword{ban}
\keyword{ban,}
\keyword{ghost}
\keyword{search}
\keyword{shadowban,}
\keyword{twitter,}
