% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtweetxtras_functions.R
\name{check_shadowban_list}
\alias{check_shadowban_list}
\title{User list shadowban test - wrapper for rtweetXtras::check_shadowban()}
\usage{
check_shadowban_list(user_list, timezone = "UTC")
}
\arguments{
\item{user_list}{A list of twitter screen_names (case insensitive)}

\item{timezone}{Optional, a timezone for the timestamp of the ban check. Default is "UTC"}
}
\description{
Checks a list of screen_names for temporary search or reply visibility reduction.
}
\examples{
listcheck <- check_shadowban_list(c("jack","jill","bob"), "Africa/Cairo")
}
\keyword{ban}
\keyword{ban,}
\keyword{ghost}
\keyword{search}
\keyword{shadowban,}
\keyword{twitter,}
