% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtweetxtras_functions.R
\name{common_follower_analysis}
\alias{common_follower_analysis}
\title{Twitter visualization of common followers}
\usage{
common_follower_analysis(user_list, follower_depth = 200, no_of_sets = 7)
}
\arguments{
\item{user_list}{A list of user_names}

\item{follower_depth}{The number of most recent followers to graph. Defaults to 200}

\item{no_of_sets}{<- The number of sets in the UpSetR graph. Defaults to 7.}
}
\description{
This function creates an UpSetR graph of common followers
     Code cribbed from Bob Rudis' 21 Recipes for Mining Twitter with Rtweet
     https://rud.is/books/21-recipes/visualizing-intersecting-follower-sets-with-upsetr.html
}
\examples{
users <- c("POTUS", "jack", "elonmusk")
common_follower_analysis(user_list = users, follower_depth = 300, no_of_sets = 7)
}
\keyword{analysis}
\keyword{follower}
\keyword{rtweet,}
\keyword{twitter,}
\keyword{visualization,}
