% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtweetxtras_functions.R
\name{common_friend_matrix}
\alias{common_friend_matrix}
\title{Twitter matrix of friend connection
Code modded from boB Rudis' 21 Recipes for Mining Twitter with Rtweet
https://rud.is/books/21-recipes/visualizing-intersecting-follower-sets-with-upsetr.html}
\usage{
common_friend_matrix(user_list, friend_depth = 200, token = NULL)
}
\arguments{
\item{user_list}{A list of user_names}

\item{friend_depth}{The number of most recent friends to include. Defaults to 200}

\item{token}{A twitter oauth token. Default is NULL and will use token in Environment}
}
\description{
This function creates a matrix of friends of a list of twitter users,
sums the number of common friends, and then ranks them in descending order.
}
\examples{
users <- c("POTUS", "jack", "elonmusk")
common_friend_matrix(user_list = users, friend_depth = 300, token = NULL)
}
\keyword{analysis}
\keyword{friend}
\keyword{rtweet,}
\keyword{twitter,}
\keyword{visualization,}
