% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createGEXF.R
\name{create_gexf}
\alias{create_gexf}
\title{Create gexf file of mentions network for export to Gephi from rtweet dataframe}
\usage{
create_gexf(
  tweet_df,
  filepath,
  include_edge_att = FALSE,
  edge_type = "directed"
)
}
\arguments{
\item{tweet_df}{Required. An rtweet dataframe of tweets}

\item{filepath}{Required. A full filepath with file name. Extension of file should be .gexf}

\item{include_edge_att}{Include edge att (status ID and Date_time). Default is FALSE to minimise Gephi processing of large networks.}

\item{edge_type}{"directed", "undirected" or "mutual". default is "directed"
#'@keywords twitter, rtweet, Gephi, gexf, sna}
}
\description{
This function creates a gexf network file of a mentions network for use with Gephi software
requires an rtweet dataframe as returned by get_timeline() or search_tweets()
Requires {rgexf} which may not be available on CRAN.
Use devtools::install_github("gvegayon/rgexf") to install
}
\examples{
rstats <- rtweet::search_tweets("#rstats", n=100, token = NULL)
create_gexf(rstats, "~/rstats.gexf", include_edge_att = FALSE, edge_type = "undirected")
}
