% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtweetxtras_functions.R
\name{get_perspective}
\alias{get_perspective}
\title{Query Google Perspective API with text}
\usage{
get_perspective(text, api_key)
}
\arguments{
\item{text}{A character string of text. Maximum size 1MB.}

\item{api_key}{A Perspective API key. Required.}
}
\description{
Queries the Google Perespective API and returns a tibble
of various probability measures
}
\examples{
df <- get_perspective("Hello, How are you?", api_key = YourAPIkeyhere)
}
\keyword{Toxicity,}
\keyword{analysis}
\keyword{text}
