% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtweetxtras_functions.R
\name{got_search}
\alias{got_search}
\title{Wrapper to undertake historical searches calling Python GetOldTweets3 library}
\usage{
got_search(
  search_string,
  since_date,
  until_date,
  n = 100,
  output_file_name,
  token = NULL,
  token_list = NULL
)
}
\arguments{
\item{search_string}{A search string (in quotes)}

\item{since_date}{Start date for the search}

\item{until_date}{Latest date for the search (NB: search works backward)}

\item{n}{Maximum number of results}

\item{output_file_name}{temporary name of GOT3 csv file - date will be appended.}

\item{token}{An OAuth token loaded in the environment for twitter's Standard API (REQUIRED if a list of tokens is not supplied)}

\item{token_list}{A list of OAuth tokens loaded in the environment (REQUIRED if token is not specified)}
}
\description{
Uses status_id to request full {rtweet} tibble of tweet data
}
\examples{
test <- got_search("Trump", since_date = "2016-09-06", until_date = "2016-11-06", n = 1000, output_file_name = "test_", token = my_token)
}
\keyword{getoldtweets3,}
\keyword{rtweet}
\keyword{twitter,}
