% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hashtagcloud.R
\name{hashtagcloud}
\alias{hashtagcloud}
\title{Hashtag wordcloud for rtweet package}
\usage{
hashtagcloud(rtweet_timeline_df, num_words = 200)
}
\arguments{
\item{rtweet_timeline_df}{A dataframe/tibble of tweets created by rtweet package through twitter API call.}

\item{num_words}{Maximum number of highest frequency terms to use in wordcloud, defaults to 200.}
}
\description{
This function creates a wordcloud (defaults to 200 terms) from terms in the "hashtags" column of an rtweet tibble of tweets
}
\examples{
df <- rtweet::get_timelines("jack", n=3200)
hashtagcloud(df, num_words = 100)
}
\keyword{rtweet,}
\keyword{twitter,}
\keyword{visualization,}
\keyword{wordcloud}
