% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtweetxtras_functions.R
\name{number_followers}
\alias{number_followers}
\title{Number followers and indicate earliest following date}
\usage{
number_followers(follower_df)
}
\arguments{
\item{follower_df}{Required. An rtweet dataframe of followers.}
}
\description{
This function consolidates the process of numbering followers and indicating earliest following date.
WARNING: Obviously not accurate for dates prior to the creation of the target account - manual adjustment required.
requires an rtweet dataframe as returned by a combination of get_followers() and lookup_users().
}
\examples{
jacks_followers <- rtweet::lookup_users(unlist(rtweet::get_followers("jack", n=5000)[,"user_id"]))
jacks_followers_numbered <- number_followers(jacks_followers)
}
\keyword{rtweet}
\keyword{twitter,}
