% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtweetxtras_functions.R
\name{prepare_got}
\alias{prepare_got}
\title{Run once function to setup call to Python Library for GetOldTweets3}
\usage{
prepare_got()
}
\description{
Ensures that {reticulate} library in stalled and GOT3 library is installed
}
\examples{
prepare_got()
}
\keyword{getoldtweets3,}
\keyword{twitter,}
