% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtweetxtras_functions.R
\name{profilecloud}
\alias{profilecloud}
\title{Twitter profile wordcloud for rtweet package}
\usage{
profilecloud(rtweet_timeline_df, num_words = 200)
}
\arguments{
\item{rtweet_timeline_df}{A dataframe/tibble of tweets or users details created by rtweet package through twitter API call.}

\item{num_words}{Maximum number of highest frequency terms to use in wordcloud, defaults to 200.}
}
\description{
This function creates a wordcloud (defaults to 200 terms) from terms in the "description" column of an rtweet tibble of tweets
}
\examples{
df <- rtweet::search_tweets("#rstats", n=3200)
profilecloud(df, num_words = 100)
}
\keyword{rtweet,}
\keyword{twitter,}
\keyword{visualization,}
\keyword{wordcloud}
