% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtweetxtras_functions.R
\name{rehydrate_got3_statuses}
\alias{rehydrate_got3_statuses}
\title{Function to rehydrate tweets into rtweet tibbles from GOT3 data format}
\usage{
rehydrate_got3_statuses(got3_df, token = NULL, token_list = NULL)
}
\arguments{
\item{got3_df}{A dataframe as returned by GetOldTweets3}

\item{token}{An OAuth token loaded in the environment for twitter's Standard API (REQUIRED if a list of tokens is not supplied)}

\item{token_list}{A list of OAuth tokens loaded in the environment (REQUIRED if token is not specified)}
}
\description{
Uses status_id to request full {rtweet} tibble of tweet data
}
\keyword{getoldtweets3,}
\keyword{rtweet}
\keyword{twitter,}
