% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtweetnet.R
\name{save_csv_edgelist}
\alias{save_csv_edgelist}
\title{Save an Edgelist csv from an Igraph Object}
\usage{
save_csv_edgelist(igraphobject, path)
}
\arguments{
\item{igraphobject}{An igraph network object}

\item{path}{Path and file name, in quotes (Filename should ideally have ".csv" extension)}
}
\description{
This function saves an igraph edgelist as a csv for export to network mapping software such as Gephi.
}
\examples{
tweets <- rtweet::search_tweets("#rstats", n= 100)
network <- rtweetXtras::rtweet_net(tweets, all_mentions = TRUE, from_threshold =2)
save_csv_edgelist(igraphobject = network, path = "~/edgelist_network.csv")
}
\keyword{export}
\keyword{igraph,}
\keyword{network,}
\keyword{rtweet,}
\keyword{twitter,}
\keyword{visualization,}
