% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtweetxtras_functions.R
\name{scatter_ts_interactive}
\alias{scatter_ts_interactive}
\title{Interactive timeline visualisation for rtweet dataset using Apache echarts}
\usage{
scatter_ts_interactive(df, title = "", print_immediately = TRUE)
}
\arguments{
\item{df}{Required. A dataframe or tibble created with {rtweet}}

\item{title}{Chart title (in quotes)}

\item{print_immediately}{Force the function to print the plot immediately. Default is TRUE,
set to FALSE to assign the object (an HTML widget)}
}
\description{
Interactive timeline of tweets indicating time of tweet, author, type (tweet, retweet, quote, reply),
and rate of tweets (posts per second) - posts per second is calculated as moving average over 10 tweets.
Posts per second are calculated separately for content tweets (tweets, quotes and replies) and copy tweets (retweets)
}
\examples{
df1 <- rtweet::searchtweets("micky mouse", n= 100)
scatter_ts_interactive (df1, title = "100 Micky Mouse tweets")
}
\keyword{echarts,}
\keyword{echarts4r}
\keyword{rtweet,}
\keyword{rtweetXtras,}
\keyword{twitter,}
