% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtweetxtras_functions.R
\name{snscrape_get_timeline}
\alias{snscrape_get_timeline}
\title{Wrapper to pull the timeline of a user}
\usage{
snscrape_get_timeline(
  screen_name,
  n = 100,
  file = "temp",
  token = NULL,
  delete_tempfile = TRUE
)
}
\arguments{
\item{screen_name}{AA twitter username/handle (no"@" required) or user_id (in quotes)}

\item{n}{Maximum number of results - twitter API has 90k rate-limit per 15 minutes}

\item{file}{temporary file name for snscrape URL list, timestamp will be appended.}

\item{token}{An OAuth token loaded in the environment for twitter's Standard API (if not specified, default token will be used)}

\item{delete_tempfile}{Clear temp file of statuses default = TRUE}
}
\description{
Calls Python script to pull status URLs of a user search, rtweet to rehdrate those statuses.
           See https://github.com/JustAnotherArchivist/snscrape
}
\examples{
test <- snscrape_get_timeline("Jack",  n = 1000, file = "test_")
}
\keyword{rtweet}
\keyword{snscrape,}
\keyword{twitter,}
