% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtweetxtras_functions.R
\name{snscrape_search}
\alias{snscrape_search}
\title{Wrapper to undertake historical searches calling Python snscrape library.}
\usage{
snscrape_search(
  search_string,
  since_date = NULL,
  until_date = NULL,
  n = 100,
  file = "temp",
  token = NULL,
  delete_tempfile = TRUE
)
}
\arguments{
\item{search_string}{A search string (in quotes)}

\item{since_date}{Start date for the search (iso format date in quotes)}

\item{until_date}{Latest date for the search (iso formatdate in quotes NB: search works backward)}

\item{n}{Maximum number of results - twitter API has 90k rate-limit per 15 minutes}

\item{file}{temporary file name for snscrape URL list, timestamp will be appended.}

\item{token}{An OAuth token loaded in the environment for twitter's Standard API (if not specified, default token will be used)}

\item{delete_tempfile}{Clear temp file of statuses default = TRUE}

\item{token_list}{A list of OAuth tokens loaded in the environment (REQUIRED if token is not specified)}
}
\description{
Calls Python script to pull status URLs of a search, rtweet to rehdrate those statuses.
           See https://github.com/JustAnotherArchivist/snscrape
}
\examples{
test <- snscrape_search("Trump", since_date = "2016-09-06", until_date = "2016-11-06", n = 1000, file = "test_", token = my_token)
}
\keyword{rtweet}
\keyword{snscrape,}
\keyword{twitter,}
