% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_compatible_csv.R
\name{write_csv_compatible}
\alias{write_csv_compatible}
\title{Save rtweet tibble as compatible csv}
\usage{
write_csv_compatible(rtweet_df, path_for_csv_file)
}
\arguments{
\item{rtweet_df}{A dataframe/tibble of tweets created by rtweet package through twitter API call.}

\item{path_for_csv_file}{a path and filename for output csv. Must include".csv"}
}
\description{
Saves csv file, appending "RT" and retweeted user name to text of retweets. Adds "text2" variable as backup.
}
\examples{
df <- rtweet::get_timelines("jack", n=3200)
write_csv_compatible(df, "~/jack_tweets.csv")
}
\keyword{compatibility}
\keyword{csv,}
\keyword{rtweet,}
\keyword{twitter,}
