% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtweetxtras_functions.R
\name{create_gexf}
\alias{create_gexf}
\title{Create gexf file of mentions network for export to Gephi from rtweet dataframe}
\usage{
create_gexf(tweet_df, filepath)
}
\arguments{
\item{tweet_df}{Required. An rtweet dataframe of tweets}

\item{filepath}{Required. A full filepath with file name. Extension of file should be .gexf
#'@keywords twitter, rtweet, Gephi, gexf, sna}
}
\description{
This function creates a gexf network file of a mentions network for use with Gephi software
requires an rtweet dataframe as returned by get_timeline() or search_tweets()
Requires {rgexf} which may not be available on CRAN.
Use devtools::install_github("gvegayon/rgexf") to install
}
\examples{
rstats <- rtweet::search_tweets("#rstats", n=100, token = NULL)
create_gexf(rstats, "~/rstats.gexf")
}
